add very basic tests for new loading logic
